added another tab for edit contractor
