+++ code/frontend/src/constants/tabs.js
@@ -1,5 +1,6 @@
 export const DASHBOARD = 'DASHBOARD';
 export const FILTERING = 'FILTERING';
 export const CONTRACTOR_INFO = 'CONTRACTOR_INFO';
+export const ADD_CONTRACTOR = 'ADD_CONTRACTOR';
 export const REPORTS = 'REPORTS';
 export const ADMIN_PANEL = 'ADMIN_PANEL';
